完善无线协议

+++ software/master/hal/CC1101.h
@@ -3,7 +3,7 @@
 
 #include "HalCtype.h"
 
-typedef void (* CC1101Recv_cb)(uint8_t *data, uint16_t len);
+typedef void (* CC1101Recv_cb)(uint8_t *data, uint8_t len);
 
 void CC1101SendData(uint8_t *data, uint16_t len);
 uint8_t CC1101ReadID(void);
